sort of added in molecule E. angle and torsion not working.

+++ REMC.v2/src/Residue.cpp
@@ -37,9 +37,9 @@ Residue::Residue(const Residue & r)
 
 float Residue::LJ_component(const Residue & rj, const double r, const AminoAcids & AminoAcidsData)
 {
-    float Eij(lambda * (AminoAcidsData.LJpotentials[aminoAcidIndex][rj.aminoAcidIndex] - e0));
+    double Eij(lambda * (AminoAcidsData.LJpotentials[aminoAcidIndex][rj.aminoAcidIndex] - e0));
     // sigmaij is the average atomic radius determined by the van der waal radius in kim2008
-    float sigmaij(0.5f * (vanderWaalRadius + rj.vanderWaalRadius));
+    double sigmaij(0.5f * (vanderWaalRadius + rj.vanderWaalRadius));
     //float r0 = powf(2.0f,(1.0f/6.0f));
     //float sigT = sigmaij / r ;
     double LJtmp(powf(sigmaij / r, 6.0f)); //sigT*sigT*sigT*sigT*sigT*sigT;
